refactor(iat): extract iat-utils and iat-graph-commands to satisfy S104 file-length
